Add the Node class
The Node class will be used to represent the nodes used by A* to find a path.

Currently the Node class stores its own coordinates for use on a grid. It also stores a target which will be a tuple representing the coordinates of the target location on the grid. Its parent node is stored as another Node object or as None if it is the starting node.

The Node class has two methods to generate its g and h values. These functions will likely work but may need to be rewritten if it is found something is not appropriate.